feat: Add Offers
