+++ src/booking/Orders/Orders.ts
@@ -12,7 +12,7 @@ export class Orders extends Resource {
   /**
    * Retrieves a paginated list of all orders. The results may be returned in any order.
    * You can optionally filter the results by the `awaiting_payment` state and sort by the `payment_required_by` date.
-   * @param {Object} [queryParams] - Pagination options (optional: limit, after, before, awaiting_payment, sort)
+   * @param {Object} [options] - Pagination query parameters (optional: limit, after, before) and other optional query parameters (awaiting_payment, sort)
    */
   public list = (options?: {
     queryParams?: PaginationMeta & ListParamsOrders
