+++ src/booking/index.ts
@@ -1,1 +1,2 @@
 export * from './OfferRequests'
+export * from './Offers'
